# CityPharmacyChain.Api.Client/Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CityPharmacyChain.Api.Host",
    "version": "1.0"
  },
  "paths": {
...
    "/Product": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "GET запрос по получению всех объектов класса препарат из базы данных",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductFullDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductFullDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductFullDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "POST запрос по добавлению объекта класса препарат в базу данных",
        "requestBody": {
          "description": "Объект класса препарат в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              }
            }
          }
        }
      }
    },
    "/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "GET запрос по получению объекта класса препарат из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор препарата",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "PUT запрос по модификации объекта класса препарат из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор препарата",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "Объект класса препарат в формате JSON",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "DELETE запрос по удалению существующего объекта класса препарат из базы данных по его идентификатору",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Идентификатор препарата",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Product/GetProductCountForEachPharmacy": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "GET запрос по получению коллекции объектов в формате JSON с информацией о всех аптеках, у которых есть в наличии заданный препарат, с указанием количества данного препарата в них",
        "parameters": [
          {
            "name": "productName",
            "in": "query",
            "description": "Название препарата",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductForPharmacyDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductForPharmacyDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductForPharmacyDto"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProductDto": {
        "type": "object",
        "properties": {
          "productCode": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "integer",
            "description": "Код препарата",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название препарата",
            "nullable": true
          },
          "productGroup": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Товарная группа препарата",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса препарат"
      },
      "ProductForPharmacyDto": {
        "type": "object",
        "properties": {
          "pharmacyName": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название аптеки",
            "nullable": true
          },
          "count": {
            "maximum": 1000,
            "minimum": 0,
            "type": "integer",
            "description": "Количество препарата на складе аптеки",
            "format": "int32",
            "nullable": true
          },
          "price": {
            "maximum": 100000,
            "minimum": 0,
            "type": "number",
            "description": "Цена препарата",
            "format": "double",
            "nullable": true
          },
          "productCode": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "integer",
            "description": "Код препарата",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название препарата",
            "nullable": true
          },
          "productGroup": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Товарная группа препарата",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса характеристики препаратов в определённой аптеке"
      },
      "ProductFullDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "description": "Идентификатор препарата",
            "format": "int32"
          },
          "productCode": {
            "maximum": 1000000,
            "minimum": 0,
            "type": "integer",
            "description": "Код препарата",
            "format": "int32",
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Название препарата",
            "nullable": true
          },
          "productGroup": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Товарная группа препарата",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Класс DTO для сущности класса препарат, содержащий её идентификатор"
      },
...
    }
  }
}